Moved call() test to FunctionRegistry and updated references.
